+++ packages/fluentui-rc/src/BaseInputTemplate/BaseInputTemplate.tsx
@@ -10,6 +10,7 @@ import {
   StrictRJSFSchema,
   labelValue,
 } from '@rjsf/utils';
+import { X } from 'lucide-react';
 
 const useStyles = makeStyles({
   input: {
@@ -86,6 +87,25 @@ export default function BaseInputTemplate<
         onBlur={_onBlur}
         aria-describedby={ariaDescribedByIds(id, !!schema.examples)}
       />
+      {options.allowClear && !readonly && !disabled && value && (
+        <button
+          type='button'
+          onClick={() => onChange('')}
+          aria-label='Clear input'
+          style={{
+            position: 'absolute',
+            left: '97%',
+            transform: 'translate(-30%,225%)',
+            backgroundColor: 'transparent',
+            cursor: 'pointer',
+            border: '2px solid #ccc',
+            zIndex: 1,
+            borderRadius: '50%',
+          }}
+        >
+          <X size={12} />
+        </button>
+      )}
       {Array.isArray(schema.examples) && (
         <datalist id={examplesId(id)}>
           {(schema.examples as string[])